requirements.txt
django-bootstrap-pagination == 1.7.1    
django-pagination-bootstrap  == 1.3      
django-cloudinary-storage==0.2.3   
cloudinary==1.20.0    
django-ckeditor == 5.9.0    
django-taggit==1.2.0
Pillow==6.1.0
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==2.2.3
gunicorn==20.0.4
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
whitenoise==5.0.1
python-decouple==3.3      
